requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.14.12
botocore==1.17.12
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-admin-honeypot==1.1.0
django-bouncy==0.2.8
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-ses==1.0.1
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.4
pem==20.1.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-http-client==3.2.7
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sendgrid==6.4.1
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
tests==0.7
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
